# Allocation fixes in controllers for Hanner Wanner Methods