Hi, we were not previously caching `work` and `dt_new` arrays which were required in controllers calculation, and hence being redundant allocation was happening. I have made a fix for these methods. The performance improvement is not humongous, but unnecessary allocations have been swooped away. 

Before:
```
julia> for tfinal = 1.0:2.0:30.0
           tspan = (0.0, tfinal)
           prob = ODEProblem(test!, u0, tspan)
           integrator = init(prob, ExtrapolationMidpointHairerWanner(), save_everystep = false)
           @time solve!(integrator)
       end
  1.555949 seconds (3.29 M allocations: 150.225 MiB, 4.37% gc time)
  0.000075 seconds (95 allocations: 12.891 KiB)
  0.000206 seconds (141 allocations: 19.172 KiB)
  0.000194 seconds (186 allocations: 25.469 KiB)
  0.000195 seconds (186 allocations: 25.469 KiB)
  0.000114 seconds (188 allocations: 25.672 KiB)
  0.000184 seconds (233 allocations: 31.938 KiB)
  0.000165 seconds (280 allocations: 38.391 KiB)
  0.000279 seconds (281 allocations: 38.406 KiB)
  0.000198 seconds (326 allocations: 44.703 KiB)
  0.000181 seconds (326 allocations: 44.703 KiB)
  0.000209 seconds (328 allocations: 44.906 KiB)
  0.000207 seconds (373 allocations: 51.172 KiB)
  0.000287 seconds (420 allocations: 57.625 KiB)
  0.000351 seconds (420 allocations: 57.625 KiB)
```
After:
```
julia> for tfinal = 1.0:2.0:30.0
           tspan = (0.0, tfinal)
           prob = ODEProblem(test!, u0, tspan)
           integrator = init(prob, ExtrapolationMidpointHairerWanner(), save_everystep = false)
           @time solve!(integrator)
       end
  0.156939 seconds (368.53 k allocations: 17.480 MiB)
  0.000054 seconds (77 allocations: 10.672 KiB)
  0.000073 seconds (113 allocations: 15.828 KiB)
  0.000093 seconds (150 allocations: 21.031 KiB)
  0.000214 seconds (150 allocations: 21.031 KiB)
  0.000095 seconds (152 allocations: 21.234 KiB)
  0.000192 seconds (188 allocations: 26.391 KiB)
  0.000195 seconds (226 allocations: 31.734 KiB)
  0.000197 seconds (226 allocations: 31.734 KiB)
  0.000156 seconds (263 allocations: 36.938 KiB)
  0.000192 seconds (263 allocations: 36.938 KiB)
  0.000311 seconds (265 allocations: 37.141 KiB)
  0.000217 seconds (301 allocations: 42.297 KiB)
  0.000207 seconds (339 allocations: 47.641 KiB)
  0.000250 seconds (339 allocations: 47.641 KiB)
```
And proportional changes in other methods.